client/requirements.txt
gpsd-py3
requests